1. README.md
The README.md file has been updated with new instructions on how to use the git-aicommit command, including setting environment variables and creating an alias.

2. autocommit.js

The autocommit.js file has been updated to fix a bug with the way commit messages are generated.

3. package.json

The package.json file has been updated to reflect the new version of the autocommit.js file.