Use a prompt to generate more human commit messages
The old version generates a **list** (1,2,3,...) of detailed changes. This is often very unhelpful, because the Changes page has a much better view. What you really want is a short, humanly-like written summary of what has changed.

This new prompt from my testing tends to write a much more human title and commit message, which is a significant step-up of usability in real projects.