Allow user to request being asked to regenerate
This add a 3rd option, "ask", in addition to "1/true" and "0/false", where the user will be asked whether they are happy with the current message, or wishes to regenerate. This increases user-friendliness, because allows you to still be lazy by not having to write commit messages, but also catch the 1% of the time where OpenAI will throw something weird or malformed.